feat: zeroDetector
The signal from the zeroDetector circuit was input to the STM32.

Each pulse of the detector flips the state of the "signal" pin. The expected waveform of "signal" would be a square wave of 60Hz, as the zeroDetector pulse has a 120Hz and it takes two of them for a period of "signal". Its pin has been analysed, and the expected waveform, verified.

The "signal" pin measurement and circuit of the zeroDetector are availiable in "assets\zeroDetector test".