feat: triggerPulse
In the softstarter, a pulse should occur in sync with the zeroDetector, but with phase/delay.

This simulates a zero detector signal through TIM10's PWM channel, which triggers TIM1's output compare to start.

The circuit schematics and scope screenshot (results) are attached in "assets/triggerPulse test".

The initial CCR2 value of TIM1 determines the phase between the zero detector and the trigger pulse, thus controling the average voltage. In this example, it's set to 1.43ms (10000 ticks) after zero detector's rising edge.

The difference between the second and initial CCR2 value sets the trigger pulse width. In this example, it's set to 1ms (7000 ticks).