refactor: remove "getTrigger()" parameter
Stop using recursion with the Newton's method. Now it's a loop.

There are no more x0 and x1. x1 overwrites x0 (now called by their final name - "triggerAngle") at each iteration.

triggerAngle starts as the initial guess.

This looks cleaner and removes the need of a guess parameter in getTrigger, which was really unnecessary and anoying.